Rewrite Tibiawiki converter (#25)
Input Side:
- Switch to WikiMedia api (massive bandwidth savings)
- Created a generic WikiMedia Template parser (its kinda janky) which removes the majority of custom regex 
- Support the new ability list template, some monsters are generated with most abilities generated for you
- Support parsing summons for creature that use the summon template
- Smarter loot parsing with fallback to the loot item template when loot statistics aren't available
- Improved runAt parsing fix #24 

Out of the 1609 active creatures 1604 are parsed from tibiawiki. The remaining 5 fail to parser due to the poor implementation of the template parser.

Output Side:
- Loot is included in the output
- Hidehealth information is included

+++ app/MonsterConverterInterface/MonsterTypes/Loot.cs
@@ -8,8 +8,8 @@ namespace MonsterConverterInterface.MonsterTypes
     {
         public string Item { get; set; }
         public decimal Chance { get; set; }
-        public decimal Count { get; set; }
-        public decimal SubType { get; set; } // Fluids in containers or rune charges
+        public int Count { get; set; }
+        public int SubType { get; set; } // Fluids in containers or rune charges
         public int ActionId { get; set; }
         public string Text { get; set; } // Sets text on writables like a letter
